Server response unit tests.  Practice with module.exports and using those functions in the unit tests
