--- app/add.js
@@ -1,5 +1,0 @@
-var add = (x,y) => {
-    return x+y;
-}
-
-module.exports = add;
