+++ app/hello_world.js
@@ -1,18 +1,11 @@
-var express = require("express");
+module.exports=  {
 
-var app = express();
+    
+    Mult: (x,y) => {
+        return x*y
+    },
 
-app.get("/", function(req, res) {
-    res.send("<h1>Hello orld</h1>");
-});
-
-app.get("/jasmine", (req, res) => {
-    res.status(200).send("<p>300</p>")
-   
-});
-
-var server = app.listen(3000,() => {
-    console.log("Listening on 3000")
-})
-
-module.exports = server;
+    Add: (x,y) => {
+        return x+y
+    }
+}
